fix(Product): upvote reviews

+++ app/Http/Controllers/Customer/ReviewController.php
@@ -30,16 +30,34 @@ class ReviewController extends Controller
 
     public function markAsUseful(Review $review)
     {
+        $upvote = $review->upvotes->where('user_id', Auth::user()->id)->first();
+        if ($upvote) {
+            return response()->json([
+                'upvotes' => count($review->upvotes),
+            ]);
+        }
+
         Upvote::create([
             'user_id' => Auth::user()->id,
             'review_id' => $review->id,
         ]);
 
-        return count($review->upvotes);
+        // return count($review->upvotes);
+
+        return response()->json([
+            'upvotes' => count($review->upvotes),
+        ]);
     }
 
     public function markAsNotUseful(Review $review)
     {
+        $upvote = $review->upvotes->where('user_id', Auth::user()->id)->first();
+        if (!$upvote) {
+            return response()->json([
+                  'upvotes' => count($review->upvotes),
+            ]);
+        }
+
         $upvote = $review->upvotes->where('user_id', Auth::user()->id)->first();
         $upvote?->delete();
